# DAML Engine: Speed up foldr